chore: update README

+++ src/arg.h
@@ -69,7 +69,7 @@ class arg_parser {
     void parse_command_args();
 
     private:
-    size_t index;
+    long unsigned int index;
     int argc;
     string bin_name;
     char **argv;
